Begin removing TaskRepository's CreateTask's menu prompts to TaskController. Add try catch, Add OperationResult record to handle error messages. Update ITaskRepository accordingly. Create shared folder and move ITaskRepository to shared interface folder and OperationResults record to Shared folder.
