+++ Shared/Interfaces/ITaskRepository.cs
@@ -1,13 +1,13 @@
 using ToDoApp.Model;
 
-namespace ToDoApp.Repository;
+namespace ToDoApp.Shared.Interfaces;
 
 public interface ITaskRepository
 {
     List<ToDoItem> GetAllTasks();
     List<ToDoItem> GetIncompleteTasks();
     List<ToDoItem> GetCompletedTasks();
-    void CreateTask(string description);
+    OperationResult CreateTask(string description);
     bool MarkTaskAsDone(int index);
     void DeleteTask(int index);
     void UpdateTask(int index, string newDescription);
