+++ Shared/OperationResult.cs
@@ -1,0 +1,3 @@
+namespace ToDoApp.Shared;
+
+public record OperationResult(bool success, string? ErrorMessage = null);
